requirements.txt
jieba
requests
lxml
zhon
pyyaml
Pillow
pyzbar
qrcode